requirements-test.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-property==0.2.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.5
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
casbin==1.43.0
catalogue==2.0.10
catboost==1.2.8
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
cloudpickle==3.1.1
cmaes==0.12.0
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.5
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==3.24.0
cymem==2.0.11
databricks-sdk==0.68.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
ecdsa==0.19.1
edgar-tool==2.1.2
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.12.4
ffmpeg-python==0.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.1
hvac==2.3.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kiwisolver==1.4.9
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.12.0
lightgbm==4.6.0
lightning==2.5.5
lightning-utilities==0.15.2
llama_stack==0.2.10.1
llama_stack_client==0.2.10
llvmlite==0.45.1
lxml==6.0.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mcp==1.18.0
mdurl==0.1.2
mlflow==3.5.0
mlflow-skinny==3.5.0
mlflow-tracing==3.5.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.12
murmurhash==1.0.13
narwhals==2.8.0
networkx==3.5
numba==0.62.1
numpy==2.3.3
openai==1.86.0
openai-whisper==20250625
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==3.3.0
packaging==25.0
pandas==2.3.0
parse==1.20.2
passlib==1.7.4
pathable==0.4.4
peewee==3.18.2
pillow==11.2.1
platformdirs==4.3.8
plotly==6.3.1
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyaml==25.5.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pydub-stubs==0.25.1.6
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.11.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-keycloak==5.7.0
python-magic==0.4.27
python-multipart==0.0.20
pytorch-forecasting==1.5.0
pytorch-lightning==2.1.0
pytz==2025.2
# pywin32==311  # Windows only - not needed in Linux container
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.3
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-rst==1.3.2
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
sec-edgar==0.0.2
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
simple-salesforce==1.12.6
simpleeval==1.0.3
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==3.0.2
starlette==0.46.2
structlog==25.4.0
sympy==1.14.0
ta==0.11.0
tenacity==8.5.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.54.0
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
waitress==3.0.2
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
xgboost==3.0.5
xmltodict==0.13.0
yarl==1.20.1
yfinance==0.2.66
zeep==4.3.1
zipp==3.23.0
